Adding a Dockerfile and helper scripts for tests
Now we can propperly test python-mode in an environment with a very well
defined python version and a vim built/compiled for this specific
version as well.

The best way is to use the `build_docker_images.sh` file in order to
create the docker image with all python versions you desire.

Also, in order to run the container you just have to do:

docker run --rm -it python-mode:<version>

Related to #977